refactor: test setup initialization
