+++ tsconfig.json
@@ -22,6 +22,6 @@
     "strictBindCallApply": false,
     "noFallthroughCasesInSwitch": false
   },
-  "include": ["src", "src/setup-application.ts"],
+  "include": ["src", "src/setup-application.ts", "test"],
   "exclude": ["dist", "node_modules"]
 }
